requirements.txt
fastapi
uvicorn[standard]
httpx
azure-identity
azure-mgmt-netapp
pytest
pytest-respx
